studyagent-backend/requirements.txt
annotated-types==0.7.0
anthropic==0.67.0
anyio==4.10.0
blinker==1.9.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
click-default-group==1.2.4
cryptography==45.0.7
distro==1.9.0
Flask==3.1.2
flask-cors==6.0.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
MarkupSafe==3.0.2
packaging==25.0
pdf2image==1.17.0
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
pluggy==1.6.0
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
pypdfium2==4.30.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
six==1.17.0
sniffio==1.3.1
sqlite-fts4==1.0.3
sqlite-utils==3.38
tabulate==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
Werkzeug==3.1.3
